requirements.txt
torch
torchvision
monai
numpy
nibabel
pydicom
PILlow
napari[pyqt6, optional]
matplotlib
open-clip-torch